BBitmap: Use 64-bit-safe types for size and offset values.
This reorders the data fields in order to make use of padding
for the additional bytes on 64-bit. I checked and the class size
should still be exactly the same on both x86 and x86_64.

Should help with or outright fix #20286.

+++ src/apps/deskbar/BarApp.cpp
@@ -1097,7 +1097,7 @@ TBarApp::_CacheTeamIcon(BarTeamInfo* barInfo, int32 size)
 	// fill with transparent
 	uint8* iconBits = (uint8*)icon->Bits();
 	if (icon->ColorSpace() == B_RGBA32) {
-		int32 i = 0;
+		size_t i = 0;
 		while (i < icon->BitsLength()) {
 			iconBits[i++] = B_TRANSPARENT_32_BIT.red;
 			iconBits[i++] = B_TRANSPARENT_32_BIT.green;
@@ -1106,7 +1106,7 @@ TBarApp::_CacheTeamIcon(BarTeamInfo* barInfo, int32 size)
 		}
 	} else {
 		// Assume B_CMAP8
-		for (int32 i = 0; i < icon->BitsLength(); i++)
+		for (size_t i = 0; i < icon->BitsLength(); i++)
 			iconBits[i] = B_TRANSPARENT_MAGIC_CMAP8;
 	}
